cmd/github/requirements.txt
click      >= 6.0
github3.py >= 0.9
pygithub3  >= 0.5
pyyaml     >= 3.0
